Added an option to format for use in TopHat questions
Added a toggle that enables TopHat formatting, which uses long narrow pages for questions, definitions and theorems, and disables all other content. The long pages can then be trimmed down and converted to image files outside of tex for use in TopHat. A handfull of the problems were formatted slightly to better fit (they check the tophat toggle for this formatting)