A more precise definition of X?
The current definition allows for X to be an arbitrary set, without restrictions on its contents. It could be a set of people, in which case the definition wouldn't make sense.

So I propose to define X as a set of vectors, or a set of points.

Might it also be a good idea to be even more precise, and define X as a set of vectors that is a subset of R^n, and then define vector v as an element of R^n? Because I'm not sure if a projection of a vector of a certain dimension onto a set of a different dimension would make sense.